spirin 5

+++ Tyuiu.KimmelDS.Sprint0.Task4.V0.Lib/DataService.cs
@@ -9,6 +9,6 @@ namespace Tyuiu.KimmelDS.Sprint0.Task4.V0.Lib
         public static int Addition(int a, int b) => a + b;
         public static int Subtraction(int a, int b) => a - b;
         public static int Multiplication(int a, int b) => a * b;
-        public static int Division(int a, int b) => a / b;
+        public static double Division(int a, int b) => a / b;
     }
 }
